Added record command + fixed console error where it checked for a message's guild in a dm

+++ commands/ratelimithandler.js
@@ -11,26 +11,28 @@ module.exports = class RateLimitHandler extends Command {
 
   async invoke(client, message) {
 
-    const guildID = message.channel.guild.id;
-    const ratelimitdb = new this._db("ratelimit");
-    const data = await ratelimitdb.get(guildID);
-    const currentsec = Math.floor(Date.now() / 1000);
-    const ratelimit = 86400; // seconds 
+    if (message.guild) {
+      const guildID = message.channel.guild.id;
+      const ratelimitdb = new this._db("ratelimit");
+      const data = await ratelimitdb.get(guildID);
+      const currentsec = Math.floor(Date.now() / 1000);
+      const ratelimit = 86400; // seconds 
 
-    // Setting up database preconditions
-    if (!(await ratelimitdb.contains(guildID))) {
-      await ratelimitdb.set(guildID, {});
-    }
+      // Setting up database preconditions
+      if (!(await ratelimitdb.contains(guildID))) {
+        await ratelimitdb.set(guildID, {});
+      }
 
-    if (data && message.author.id !== client.user.id) {
-      if (data[message.author.id] != undefined) {
+      if (data && message.author.id !== client.user.id) {
+        if (data[message.author.id] != undefined) {
 
-        if (currentsec - data[message.author.id] < ratelimit) {
-          message.delete();
+          if (currentsec - data[message.author.id] < ratelimit) {
+            message.delete();
 
-        } else {
-          data[message.author.id] = currentsec;
-          await ratelimitdb.set(guildID, data);
+          } else {
+            data[message.author.id] = currentsec;
+            await ratelimitdb.set(guildID, data);
+          }
         }
       }
     }
